Extracted ProductList component
